requirements.txt
# Legal Guardian MVP — unified stack
streamlit>=1.28.0
pandas>=2.0.0
python-dotenv>=1.0.0
groq>=0.4.0
faiss-cpu>=1.7.4
numpy>=1.24.0
sentence-transformers>=2.2.0
# torch/torchvision: sentence-transformers needs torch; torchvision satisfies optional
# transformers imports used by Streamlit's file watcher. Pin torchvision to your torch build if needed.
torchvision>=0.15.0
# CPU-only PyTorch (optional, smaller): pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
langchain-core>=0.2.0
langchain-text-splitters>=0.2.0
langchain-community>=0.2.0
pypdf>=4.0.0
docx2txt>=0.8
google-api-python-client>=2.100.0
google-auth>=2.22.0
duckduckgo-search>=6.0.0
